Update: move CStyleCommentBlockFilter to block filter module
